Add Draft.js to the write page #65
Took a rich editor example and added it, supporting:
- h1-h6
- ul & ol
- em, b, blockquote, code
- pre, code
- underlined

It tries to render the content as markdown and submits it as the body.
Also stores serialized Draft.js state on the `json_metadata`.

On development an example Draft.js state that could be loaded from
`json_metadata` is loaded from a JSON file.