Correct Parent Node Bug
As seen in https://i.imgur.com/Y9irhyK.png when someone puts an iframe into the editor without a parent element (such as a div) there is an error. This is because there is no parentNode when it is setup this way. So lines 146-151 make sure if there is no parentNode, then it will create a wrapping div. The problem lies at 158. Without the parentNode you can't replace the child. It took me five hours to find this out.

Line 139 allows for all iframes to have an embedded thumbnail.