Remove border-left/right and border-radius in Wallet/Activity page
In smaller device, page takes full width so blocks in Wallet and
Activity page now longer have border-left and border-right. These
CSS properties are visible for medium device and above.

Same for border-radius, on smaller device it is creating unwanted
white-space. Border-radius is only available on medium device and
above.