fix-reblogged-by (#2150)
post.first_reblogged_by returns null.

should use post.reblogged_by

https://github.com/steemit/condenser/blob/7430aca58ae9f86e501f52e3cf8dba5efc9fe3e0/src/app/components/cards/PostSummary.jsx#L70

In addition, to show 'reblogged' in blog (not feed), the current code should be changed. Currently there is no easy way to differentiate if it's feed or blog. But any resteemed post in a blog is easily noticeable by author's name, so not important as reblogged_by in feed.